:arrow_up: #255 storybook のバージョンを最新安定版に更新
