--- .storybook/main.js
@@ -1,14 +1,0 @@
-module.exports = {
-  stories: ['../src/**/*.stories.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
-  addons: [
-    '@storybook/addon-links',
-    '@storybook/addon-essentials',
-    '@storybook/addon-interactions',
-    '@storybook/addon-a11y',
-    'storybook-addon-next',
-  ],
-  framework: '@storybook/react',
-  core: {
-    builder: 'webpack5',
-  },
-};
