+++ package.json
@@ -19,8 +19,8 @@
     "test:coverage": "jest --collect-coverage",
     "test:ci": "jest --runInBand --collect-coverage",
     "generate:css-types": "hcm 'src/**/*.module.{css,scss}'",
-    "storybook": "start-storybook -s ./public -p 6008",
-    "build-storybook": "build-storybook -s public",
+    "storybook": "storybook dev -p 6008 -s ./public",
+    "build-storybook": "storybook build",
     "chromatic": "chromatic --project-token=$CHROMATIC_PROJECT_TOKEN"
   },
   "dependencies": {
@@ -34,15 +34,14 @@
     "ress": "^5.0.2"
   },
   "devDependencies": {
-    "@storybook/addon-a11y": "^6.5.16",
-    "@storybook/addon-actions": "^6.5.16",
-    "@storybook/addon-essentials": "^6.5.16",
-    "@storybook/addon-interactions": "^6.5.16",
-    "@storybook/addon-links": "^6.5.16",
-    "@storybook/builder-webpack5": "^6.5.16",
-    "@storybook/manager-webpack5": "^6.5.16",
-    "@storybook/react": "^6.5.16",
-    "@storybook/testing-library": "^0.0.13",
+    "@storybook/addon-essentials": "^7.2.3",
+    "@storybook/addon-interactions": "^7.2.3",
+    "@storybook/addon-links": "^7.2.3",
+    "@storybook/addon-onboarding": "^1.0.8",
+    "@storybook/blocks": "^7.2.3",
+    "@storybook/nextjs": "^7.2.3",
+    "@storybook/react": "^7.2.3",
+    "@storybook/testing-library": "^0.2.0",
     "@testing-library/react": "^14.0.0",
     "@testing-library/react-hooks": "^8.0.1",
     "@types/eslint": "^8.21.3",
@@ -71,10 +70,8 @@
     "msw-storybook-addon": "^1.8.0",
     "npm-run-all": "^4.1.5",
     "postcss": "^8.4.21",
-    "postcss-jsx": "^0.36.4",
-    "postcss-syntax": "^0.36.2",
     "prettier": "^2.8.6",
-    "storybook-addon-next": "^1.7.3",
+    "storybook": "^7.2.3",
     "stylelint": "^15.6.2",
     "stylelint-config-prettier": "^9.0.5",
     "stylelint-config-recess-order": "^4.0.0",
@@ -82,7 +79,6 @@
     "stylelint-order": "^6.0.3",
     "ts-jest": "^29.0.5",
     "typescript": "4.9.5",
-    "webpack": "^5.76.3",
     "whatwg-fetch": "^3.6.2"
   },
   "msw": {
